feat: EventStageManager.jsの分割とコンポーネント化
- SeasonalEventManager.js: 季節イベント管理（1,077語）
- EventNotificationSystem.js: 通知システム（1,394語）
- EventHistoryManager.js: 履歴管理（1,658語）
- EventRankingSystem.js: ランキングシステム（1,947語）
- EventStageManager.js: 9,914語 → 933語（90%削減）

新アーキテクチャによる保守性・拡張性の大幅向上
レガシーAPIとの互換性維持
